chore: refactor cam db (#20)
