+++ src/cam-database/brands.ts
@@ -1,0 +1,12 @@
+import { Brand } from './types';
+
+const bd: Brand = {
+    id: 'bd',
+    name: 'Black Diamond',
+};
+const mt: Brand = {
+    id: 'mt',
+    name: 'Metolius',
+};
+
+export { bd, mt };
