+++ src/cam-database/models.ts
@@ -1,0 +1,20 @@
+import { Model } from './types';
+import { bd, mt } from './brands';
+
+const c4: Model = {
+    id: 'c4',
+    name: 'Camalot C4 (2019)',
+    brand: bd,
+};
+const c3: Model = {
+    id: 'c3',
+    name: 'Camalot C3',
+    brand: bd,
+};
+const mc: Model = {
+    id: 'mc',
+    name: 'Ultralight Master Cam',
+    brand: mt,
+};
+
+export { c4, c3, mc };
